requirements.txt
suds-community==0.8.5
